pyproject.toml
[tool.poetry]
name = "beautybusinessscraper"
version = "0.1.0"
description = ""
authors = ["dasdatasensei <jody@thedatasensei.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.11.12"
beautifulsoup4 = "^4.13.3"
openai = ">=1.0.0"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
httpx = "^0.28.1"
tenacity = "^9.0.0"
playwright = "^1.50.0"
playwright-stealth = "^1.0.6"
requests = "^2.32.3"
selenium = "^4.28.1"
ace-tools = "^0.0"
webdriver-manager = "^4.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
